bot_scheduler/requirements.txt
pandas==1.0.1
numpy==1.18.1
yfinance==0.1.63
python-dotenv==0.15.0
python-telegram-bot==13.1
psycopg2-binary==2.8.6
sqlalchemy==1.3.22
lxml==4.6.3